requirements.txt
Django>=5.2.6,<5.3
django-ninja>=1.4.3,<1.5
djangorestframework_simplejwt>=5.5.1,<5.6
